Fix formhelper to create proper fields
Previously, using the form helper to create fields for the metadata
resulted in wrong name values for the html tags and thus the values
would not end up being saved. This properly sets the name of those tags
so they can be saved again. In addition, this also adds the date type.