Fix queries starting with '('
Not sure what the issue here was, but I guess the parser got stuck in
some rule in element (I guess text? not sure). However, that doesn't
make much sense, since it should just backtrack. It's also not the fact
that the parsers reads it as `element | str('(')`, because putting
brackets around `str('(') .... srt(')')` doesn't work either. Putting it
 this way around also has another benefit as it can cancel early since
it only needs the first char.