Manually specify seek time when doing thumbnails
Using the `seek_time` option uses it as an encoding option, which is not
really what we want. It requires ffmpeg to encode the whole video up to
that point and generate an image out of it, causing really long
generation times. Instead, we specify it as an input parameter manually
so it will tell ffmpeg to do a fast seek which is _much_ faster. The
possibly lost accuracy can be ignored here since we don't care about a
perfect representation, we just need a decent thumbnail.