Add refresh button back on top
Now that we moved to file hashes, refreshing the library is much faster
and it makes much more sense to have this button now. For now it only
runs the job but doesn't refresh or update the view. This is mostly
because refreshing the view for a large library still laggs the browser
quite hard.