Use img-2 for loading images
This should improve performance on initial load of bigger video lists
when having a big for example. The styling doesn't quite work out
right yet, as it seems to always be a few pixel off. Another issue I
have with it is that I have to specify the exact size of the image,
but we have to live with that for now.