Improve thumbnail caching and remove img-2
Img2 has quite a few issues that I ran into when using it in an bigger
library, such as screwing with the sizes, not properly loading and
overwriting a lot of my styling. Instead I thought it would be more
worthwhile to properly cache thumbnails, thus I changed the thumbnails
to no longer contain their full path and they will be requested via name.
Thus we can set aggressive caching on them as the names will be unique.