Automatically save video upon handling metadata
Right now the handlers need to call save themselves which is not very
optimal. Videos may get saved more than they need to be, or not at all
if the handler forgets. To make this process a little easier, the
handler manager will now save videos after calling events on them
automatically.