Fix exception when trying to parse bracketed or expression
When using a bracketed expression, the parser returned the content as a
hash with a key 'query', which should only happen for the root rule.
This was because we reused the root rule inside the bracketed
expression. To fix this we only use the same rule content, but without
telling the parser a name for it.