Fix broken query execution
Right now if you did a query that was more complex than a simple
property check, it would not have worked. For some reason, mongoid just
overrides previous instances of the same attribute even when using
`union` or `intersect`. Thus a query like this:

    Video.where(:type => 'a').union.where(:type => 'b')

is equivalent to:

    Video.where(:type => 'b')

To fix this, we have to do a little workaround. Instead of using `union`
and `intersect`, we use internal selectors to build the query. For
example, the above would look like this:

    Video.or(Video.where(:type => 'a').selector,
        Video.where(:type => 'b').selector)

Obviously, in these examples you could merge them together, however, we
can't easily do that in our query parser (it's not technically
impossible, it's just much more complex). This solution works just fine
in our use case and thus was changed to this behavior.