Properly cache video cards
Initial render is (for my collection) at around 8s which is not great.
Rendering each video card doesn't take much time individually (~4ms),
but for such a big library it stacks up to that number. Since we only
need to re-render a card when the video changed, we can easily cache
each card. To do that we first wrap the partial in a cache block and
then use a collection renderer which we tell about the caching. This
puts the rendering down to less than a second for me for subsequent
site loads.

One has to note that this caching only comes into effect when running in
production mode or having a file called `caching-dev.txt` in the `tmp`
project dir.