Adjust styling of select2
This extends and changes much of the styling of the current select2.
This was done to adjust to the dark theme since it was still completely
white and also now unreadable. As an improvement we also now use the
same style for the tags here as we do in the view page.