Add feature to edit multiple videos simultaneously
This allows a user to select multiple videos and edit metadata on all of
 them in one go. The design is currently a little lacking but works as a
 first iteration.

It would also be better to not do the redirect in the server, it should
just do a reload on the client. Moreover, the script code should be
moved somewhere else more appropriate and not be in the shared view.