Allow for negating property query
We allow negation for boolean types, but we were missing the
negation/opposite of property queries. This is now similar to how you'd
filter boolean values or empty taglists using a prefixed `-`, such as
"-time:night", which would only present you with videos with `time` not
being `night`. Note that empty values would match this query as they're
not equal.

We now have to parse the property node before the boolean node because
otherwise the boolean node parser will succeed, yet the remaining query
will fail. So instead we parse the property first, so we check the
longer chain, and only later check the shorter one.