Fix negation of properties in query not working for empty text
If you wanted to check if a certain property was empty, you could use
`abd:""`. However, since we provided a shortcut for tag lists which
looked just like a boolean query, one could expect this to work for all
other kinds, which didn't work. Previously it assumed a boolean value,
which obviously would not work with empty text, number or other types.
Instead for anything else we now check against empty values or the
opposite depending on the query.