Fix not being able to connect text queries
When trying to do a full text search query and connecting two, it would
fail (such as "hello OR test") with an error. This was due to the fact
that text queries were not properly transformed if they were not present
 at the root query level. This is always the case when you use AND or OR
 in the query as the root node will be a combination node.

To fix that, we now properly transform text queries just like we do with
 all the other kinds. This might result in somewhat slower performance
since we no longer combine nodes together, but that can be done in the
future.