Allow for reordering videos in a normal playlist
Right now when creating a playlist, there's no way to influence the
order where the videos will be in. Mongoid will always load them in the
order of IDs and thus it doesn't matter in which order they are in the
datastore. To work around that, we have a separate accessor for
accessing them in the order they are defined.

Ordering in the UI currently works by dragging videos back and forth. It
 doesn't look perfect but it works.