Show proper title in pages
For now this only used a static name for each page. This is not really
helpful once multiple tabs are open and you're string to see what tab
shows what. This now makes it a little bit more clear what the tab is
showing by providing a little bit more information about the current
page.