Go back to MIT license
I've started this project wanting to use MIT as a license mostly because
 I just wanted the project or rather the resulting software to exist and
 that I can use it. Thus I just wanted a simple license that allows
people to do what they want with it. This over time evolved into me
wanting that if anyone published their custom version had to publish the
 changed source too so I could the patches into vima too. But I no
longer think that I care about that enough and would rather have a
simpler license that just gets out of the way. Therefor MIT it is again.