Automatically include fields on changes
When editing multiple videos, after changing a value, the field should
automatically be included. Right now, one needs to manually check the
include checkbox. To ease this workflow the checkbox will not
automatically be checked after a non-empty value was provided or
unchecked if the given value changed to empty. The checkbox can still be
 selected manually to possibly overwrite with an empty value.