Allow for creating library at runtime and multiple library locations
Configuring everything inside a configuration file is quick and dirty
and allows for getting started fast. However, it's not really user
friendly as one would have to know which configuration setting has which
 value and what structure. This becomes increasingly difficult as the
configuration has more power. As a result I'll be slowly moving away
from the file based configuration and this is a first step towards that.

This moves the library location to a database entity which allows
editing the locations in the browser. There's now a new page for
"settings" which includes these new libraries. It is not the final
design but an working state.

To migrate an existing system, one can run the rake task
`migrate_3_create_library_store` to automatically migrate the existing
database. The existing `dir` entry in the configuration file can be
removed afterwards.