Fix thumbnails given the most recent commit
The last commit failed to address thumbnails when having multiple
libraries as there's no single library anymore where the thumbnails will
 be stored.

Now the thumbnails will be stored in the given folder in each library.
I.e. each library will contain a thumbnail folder which holds the
thumbnails for the videos of that library. This also results in a
difference in how thumbnails can be requested from the frontend.
Previously, we could just use the name of the file, check for the file
in our thumbnail directory and be done with it. Since we now have
multiple libraries this would be rather silly to do for each thumbnail
over each library. Not to mention this could be used for a path
traversal attack. Instead, we now just give the video and the thumbnail
index and it will give us the corresponding image. The downside is that
we can't easily cache anymore since we don't know when the thumbnails
have changed. We do a little caching on when the video was last updated
but that happens quite often.