Move `lib` directory into application to ease loading
By default, rails already loads most the classes that are inside the
`app` folder. We can use this for our metadata and parser to avoid
making custom configurations for them.for our handlers this doesn't
fully work because we need them to be initialized directly to make sure
they're registered. This can obviously be improved upon, but for now
that's how it is.